+++ spec/ncs_navigator/configuration_spec.rb
@@ -107,10 +107,10 @@ module NcsNavigator
         from_hash.study_center_id.should == '234'
       end
 
-      it 'is mandatory' do
+      it 'is not mandatory' do
         input_hash['Study Center'].delete 'sc_id'
         lambda { from_hash }.
-          should raise_error("Please set a value for [Study Center]: sc_id")
+          should_not raise_error
       end
     end